+++ CHANGELOG.md
@@ -22,6 +22,7 @@ ENHANCEMENTS:
 
 * Cancelling an Airlock request triggers deletion of the request container and files ([#2584](https://github.com/microsoft/AzureTRE/pull/2584))
 * Move admin-vm from core to a shared service ([#2624](https://github.com/microsoft/AzureTRE/pull/2624))
+* Remove obsolete docker environment variables ([#TBD](https://github.com/microsoft/AzureTRE/pull/TBD))
 
 BUG FIXES: